Merge branch 'develop' of https://github.com/bob1983/qiitahackason_20141213_ios into develop

Conflicts:
	QiitaMeter/Models/ModelObjects/JANPoint.h
	QiitaMeter/Models/Services/JANQiitaUserInfoService.m

+++ QiitaMeter/Models/ModelObjects/JANPoint.h
@@ -11,5 +11,5 @@
 @interface JANPoint : RLMObject
 @property NSString *qiitaId;
 @property NSInteger totalPoint;
-@property float gaugePersentValue;
+@property CGFloat gaugePersentValue;
 @end
